WIP recombination
Currently there's an error with recombining the tail - showing as no tail when there should be one.
How am I going to handle the situation when the recombination has increased the number of sections?
I could recombine the object configs first, then move to the robot?
Would a recombined new section be better than an averaged new section?